+++ addon/services/map-service.js
@@ -1,13 +1,6 @@
 import Ember from 'ember';
 import serviceMixin from '../mixins/ags-service-mixin';
+import layerMixin from '../mixins/layers';
 
-export default Ember.Service.extend(serviceMixin, {
-
-  /**
-   * Get the layer info
-   */
-  getLayerInfo (url) {
-    return this.request(url + '?f=json', {method: 'GET'});
-  }
-
+export default Ember.Service.extend(layerMixin, serviceMixin, {
 });