+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Added
+* Map Service supports getLayerInfo, getLayersInfo, getServerInfo
+* New "Vector Service" supports common features between map and feature service
+
 ## 0.1.4
 ### Changed
 - remove ember-ajax and related code